% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exercises.R
\name{get_all_exercises}
\alias{get_all_exercises}
\title{Get all exercises (paged)}
\usage{
get_all_exercises(org_id)
}
\arguments{
\item{org_id}{Numeric organization id.}
}
\value{
A data frame of exercises.
}
\description{
Get all exercises (paged)
}
\examples{
\dontrun{
  perch_auth(Sys.getenv("PERCH_TOKEN"))
  oid <- get_organization_id()
  exs <- get_all_exercises(oid)
}
}
