% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_all_users}
\alias{get_all_users}
\title{Get all users (paged)}
\usage{
get_all_users(org_id)
}
\arguments{
\item{org_id}{Numeric organization id.}
}
\value{
A data frame of users.
}
\description{
Get all users (paged)
}
\examples{
\dontrun{
  perch_auth(Sys.getenv("PERCH_TOKEN"))
  oid <- get_organization_id()
  users <- get_all_users(oid)
}
}
