% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organization_id.R
\name{get_organization_id}
\alias{get_organization_id}
\title{Get your organization's id (from /v2/user)}
\usage{
get_organization_id()
}
\value{
Numeric org id.
}
\description{
Calls the Perch API and extracts \code{org_id} from the response.
}
\examples{
\dontrun{
  perch_auth(Sys.getenv("PERCH_TOKEN"))
  get_organization_id()
}
}
