% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{perch_auth}
\alias{perch_auth}
\title{Authenticate with the Perch API}
\usage{
perch_auth(token)
}
\arguments{
\item{token}{A valid API token string.}
}
\value{
Invisibly returns TRUE.
}
\description{
Authenticate with the Perch API
}
